Remove unused Jekyll stuff from Travis config
This was a relic from when I was planning to publish the microsite via
Travis. For now I'm happy to publish it manually so I'll get rid of it.
It slows down the build and appears to have broken at some point
recently (see https://travis-ci.org/cb372/cats-retry/builds/472633596)