Zio interoperability
It order to empowered more taggeless final users and simplify migrations
from/to ZIO we introduce some interoperability library for it. This
library works the same way that Monix interoperability. It's currently
working only on JVM, ScalaJS support will follow (for now it won't compile).

Be aware that if you are completely commit to either Monix or ZIO both
libraries already have their internal support for retry policies.