POC: API changes to support adaptation
This does not compile. I've only changed the public API, not the
implementation methods. This is just a strawman proposal to aid
discussion.

The `wasSuccessful` and `onFailure` params (or similarly the
`isWorthRetrying` and `onError` params) are collapsed down to a single
`ResultHandler`. This handler inspects the result (or error, depending
on which API we are using) and details of retries so far, and can choose
to:
* stop retrying (either because the result was successful, or because
  the error was not worth retrying)
* continue retrying the same action (assuming the retry policy wants to
  keep retrying)
* adapt to the result and try a new action (again, assuming the retry
  policy allows it)

The handler can also do any logging it needs to, because the result is
monadic.