API changes to support adaptation (#534)
Redesign the API to allow users to adapt their action in response to
failures or errors, instead of stubbornly retrying the same action.

The `wasSuccessful` and `onFailure` params (or similarly the
`isWorthRetrying` and `onError` params) are collapsed down to a single
`ResultHandler`. This handler inspects the result (or error, depending
on which API we are using) and details of retries so far, and can choose
to:
* stop retrying (either because the result was successful, or because
the error was not worth retrying)
* continue retrying the same action (assuming the retry policy wants to
keep retrying)
* adapt to the result and try a new action (again, assuming the retry
policy allows it)

The handler can also do any logging it needs to, because the result is
monadic. That allows us to remove the `onFailure` and `onError` hooks,
which makes the method signatures a bit more friendly.

Thanks to the flexibility provided by the `ResultHandler`, I was also
able to reduce the number of combinators, so the API is a lot slimmer
now.

Before:
* `retryingOnFailures(policy, wasSuccessful, onFailure)`
* `retryingOnSomeErrors(policy, isWorthRetrying, onError)`
* `retryingOnAllErrors(policy, onError)`
* `retryingOnFailuresAndSomeErrors(policy, wasSuccessful,
isWorthRetrying, onFailure, onError)`
* `retryingOnFailuresAndAllErrors(policy, wasSuccessful, onFailure,
onError)`

After:
* `retryingOnFailures(policy, valueHandler)`
* `retryingOnErrors(policy, errorHandler)`
* `retryingOnFailuresAndErrors(policy, errorOrValueHandler)`

While the new API is more flexible, it also means the user needs to
write more code for simple cases like "retry on all errors" or "retry
until the result matches this predicate". I have included a couple of
helpers to handle those cases:

```scala
// retry without adaption until the result matches a predicate, similar to the old retryingOnFailures API
action.retryingOnFailures(
  policy,
  ResultHandler.retryUntilSuccessful(
    isSuccessful = _.statusCode == 200,
    log = noop // don't bother with any logging on retries
  )
)

// retry without adaption on all errors, similar to the old retryingOnAllErrors API
action.retryingOnErrors(
  policy,
  ResultHandler.retryOnAllErrors(log = noop)
)
```