Pass the action result to the retry policy
When asking the retry policy to decide the next step, pass the result of
the last attempt. Depending on the combinator, this could be a value or
an error.

This allows users to implement more exotic retry policies, e.g. choosing
to retry some kinds of error quickly, but using exponential backoff for
others.

All the built-in policies maintain their existing behaviour: they ignore
the action result and decide the next step based on static rules.

This was quite a noisy change because it involved adding a new type
parameter to `RetryPolicy`.