Remove obsolete Mergify and release-drafter configs (#556)
* Mergify
* We have branch protection, and it's a lot more powerful than it used
to be, so there's not really any reason to use Mergify any more.
* It used to be useful for auto-merging Scala Steward patch-level
updates, but branch protection blocks that. As somebody needs to come
and approve the PR, that person may as well click the button to merge it
as well.
* release-drafter
* I used to use this to build release notes, but nowadays you just click
a button in the GitHub UI to achieve the same thing
    * I've already removed the release-drafter app from the repo